Language.C.Inline.Cpp

Description

Module exposing a Context to inline C++ code. We only have used this for experiments, so use with caution. See the C++ tests to see how to build inline C++ code.

Documentation

module Language.C.Inline

cppCtx :: Context

The equivalent of baseCtx for C++. It specifies the .cpp file extension for the C file, so that g++ will decide to build C++ instead of C. See the .cabal test target for an example on how to build.

cppTypePairs :: [(CIdentifier, TypeQ)] -> Context

using :: String -> DecsQ

Emits an using directive, e.g.

C.using "namespace std" ==> using namespace std

data AbstractCppExceptionPtr

Marks an std::exception_ptr. Only used via Ptr.