Language.C.Inline.Cpp.Exception

Description

A module that contains exception-safe equivalents of inline-c QuasiQuoters.

Documentation

data CppException

An exception thrown in C++ code.

Constructors

CppStdException CppExceptionPtr ByteString (Maybe ByteString)

CppHaskellException SomeException

CppNonStdException CppExceptionPtr (Maybe ByteString)

type CppExceptionPtr = ForeignPtr AbstractCppExceptionPtr

toSomeException :: CppException -> SomeException

Like toException but unwrap CppHaskellException

throwBlock :: QuasiQuoter

Like tryBlock, but will throw unwrapped CppHaskellExceptions or other CppExceptions rather than returning them in an Either

tryBlock :: QuasiQuoter

Similar to block, but C++ exceptions will be caught and the result is (Either CppException value). The return type must be void or constructible with {}. Using this will automatically include exception, cstring and cstdlib.

catchBlock :: QuasiQuoter

Variant of throwBlock for blocks which return void.

tryBlockQuoteExp :: QuasiQuoter -> String -> Q Exp