Language.C.Inline.Cpp.Exceptions

Description

Deprecated: Language.C.Inline.Cpp.Exceptions is deprecated in favor of Language.C.Inline.Cpp.Exception which changes the CppException data type to preserve the exception for custom error handling.

Documentation

data CppException

An exception thrown in C++ code.

Constructors

CppHaskellException SomeException

pattern CppStdException :: String -> CppException

pattern CppOtherException :: Maybe String -> CppException

toSomeException :: CppException -> SomeException

Like toException but unwrap CppHaskellException

throwBlock :: QuasiQuoter

Like tryBlock, but will throw unwrapped CppHaskellExceptions or other CppExceptions rather than returning them in an Either

tryBlock :: QuasiQuoter

Similar to block, but C++ exceptions will be caught and the result is (Either CppException value). The return type must be void or constructible with {}. Using this will automatically include exception, cstring and cstdlib.

catchBlock :: QuasiQuoter

Variant of throwBlock for blocks which return void.